+++ tests/test_mctpd.py
@@ -9,7 +9,7 @@ from mctp_test_utils import (
     mctpd_mctp_endpoint_common_obj,
     mctpd_mctp_endpoint_control_obj
 )
-from mctpenv import Endpoint, MCTPSockAddr, MCTPControlCommand
+from mctpenv import Endpoint, MCTPSockAddr, MCTPControlCommand, MctpdWrapper
 
 # DBus constant symbol suffixes:
 #
@@ -763,6 +763,44 @@ async def test_del_interface_last(dbus, mctpd):
     with pytest.raises(asyncdbus.errors.DBusError):
         await mctpd_mctp_network_obj(dbus, iface.net)
 
+""" Delete an interface with peers attached, ensure all are gone """
+async def test_del_interface_with_peers(dbus, mctpd):
+    net = mctpd.system.interfaces[0].net
+    iface = mctpd.system.Interface(
+        'mctp1', 2, net,  bytes([0x10]), 68, 254, True,
+    )
+    await mctpd.system.add_interface(iface)
+
+    eps = [
+        Endpoint(iface, bytes([0x11])),
+        Endpoint(iface, bytes([0x12])),
+        Endpoint(iface, bytes([0x13])),
+    ]
+
+    mctp = await mctpd_mctp_iface_obj(dbus, iface)
+
+    paths = []
+    for ep in eps:
+        mctpd.network.add_endpoint(ep)
+        (eid, _, path, _) = await mctp.call_setup_endpoint(ep.lladdr)
+        assert eid == ep.eid
+        paths.append(path)
+
+    await mctpd.system.del_interface(iface)
+
+    # interface should be gone
+    with pytest.raises(asyncdbus.errors.DBusError):
+        await mctpd_mctp_iface_obj(dbus, iface)
+
+    # .. but the network should remain, as the default interface is still
+    # present
+    _ = await mctpd_mctp_network_obj(dbus, net)
+
+    for path in paths:
+        with pytest.raises(asyncdbus.errors.DBusError) as ex:
+            ep = await mctpd_mctp_endpoint_common_obj(dbus, path)
+        assert str(ex.value).startswith("Unknown object")
+
 """ Remove and re-add an interface """
 async def test_add_interface(dbus, mctpd):
     net = 1
@@ -908,7 +946,6 @@ async def test_endpoint_allocate_eid(dbus, mctpd):
     query_types = list(await ep.get_supported_message_types())
     assert (query_types == mctpd.network.endpoints[1].types)
 
-
 """ Test that ignore_eids parameter works correctly """
 async def test_assign_endpoint_static_with_ignore_eids(dbus, mctpd):
     iface = mctpd.system.interfaces[0]
@@ -933,3 +970,60 @@ async def test_assign_endpoint_static_with_ignore_eids(dbus, mctpd):
     assert neigh.lladdr == dev.lladdr
     assert neigh.eid == static_eid
     assert len(mctpd.system.routes) == 2
+
+""" Test that we use the minimum EID from the dynamic_eid_range config """
+async def test_config_dyn_eid_range_min(nursery, dbus, sysnet):
+    (min_dyn_eid, max_dyn_eid) = (20, 254)
+    config = f"""
+    [bus-owner]
+    dynamic_eid_range = [{min_dyn_eid}, {max_dyn_eid}]
+    """
+
+    # since we're specifying per-test config, we create the wrapper directly
+    # rather than using the fixture.
+    mctpd = MctpdWrapper(dbus, sysnet, config = config)
+    await mctpd.start_mctpd(nursery)
+
+    iface = mctpd.system.interfaces[0]
+    ep = mctpd.network.endpoints[0]
+
+    mctp = await mctpd_mctp_iface_obj(dbus, iface)
+    (eid, net, path, new) = await mctp.call_setup_endpoint(ep.lladdr)
+    assert eid == min_dyn_eid
+    assert ep.eid == eid
+
+    res = await mctpd.stop_mctpd()
+    assert res == 0
+
+""" Test that we use the maximum EID from the dynamic_eid_range config """
+async def test_config_dyn_eid_range_max(nursery, dbus, sysnet):
+    (min_dyn_eid, max_dyn_eid) = (20, 21)
+    config = f"""
+    [bus-owner]
+    dynamic_eid_range = [{min_dyn_eid}, {max_dyn_eid}]
+    """
+
+    mctpd = MctpdWrapper(dbus, sysnet, config = config)
+    await mctpd.start_mctpd(nursery)
+
+    iface = mctpd.system.interfaces[0]
+    mctp = await mctpd_mctp_iface_obj(dbus, iface)
+
+    mctpd.network.add_endpoint(Endpoint(iface, bytes([0x01]), types = [0, 1]))
+    mctpd.network.add_endpoint(Endpoint(iface, bytes([0x02]), types = [0, 1]))
+
+    for i in range(0, 2):
+        ep = mctpd.network.endpoints[i]
+        (eid, net, path, new) = await mctp.call_setup_endpoint(ep.lladdr)
+        assert eid >= 20 and eid <= 21
+
+    # we should have run out of EIDs
+    with pytest.raises(asyncdbus.errors.DBusError) as ex:
+        ep = mctpd.network.endpoints[2]
+        (eid, net, path, new) = await mctp.call_setup_endpoint(ep.lladdr)
+
+    assert str(ex.value) == "Ran out of EIDs"
+    assert mctpd.network.endpoints[2].eid == 0
+
+    res = await mctpd.stop_mctpd()
+    assert res == 0